+++ tests/Unit/Schema/Types/LaravelEnumTypeTest.php
@@ -4,10 +4,11 @@ namespace Tests\Unit\Schema\Types;
 
 use Nuwave\Lighthouse\Schema\TypeRegistry;
 use Nuwave\Lighthouse\Schema\Types\LaravelEnumType;
+use Nuwave\Lighthouse\Support\AppVersion;
 use PHPUnit\Framework\Constraint\Callback;
 use Tests\TestCase;
+use Tests\Utils\LaravelEnums\AOrB;
 use Tests\Utils\LaravelEnums\LocalizedUserType;
-use Tests\Utils\LaravelEnums\UserType;
 
 class LaravelEnumTypeTest extends TestCase
 {
@@ -20,13 +21,17 @@ class LaravelEnumTypeTest extends TestCase
     {
         parent::setUp();
 
+        if (AppVersion::atLeast(7.0)) {
+            $this->markTestSkipped('TODO remove this once bensampo/laravel-enum supports Laravel 7');
+        }
+
         $this->typeRegistry = $this->app->make(TypeRegistry::class);
     }
 
     public function testMakeEnumWithCustomName(): void
     {
         $customName = 'CustomName';
-        $enumType = new LaravelEnumType(UserType::class, $customName);
+        $enumType = new LaravelEnumType(AOrB::class, $customName);
 
         $this->assertSame($customName, $enumType->name);
     }
@@ -42,22 +47,22 @@ class LaravelEnumTypeTest extends TestCase
     {
         $this->schema = /** @lang GraphQL */ '
         type Query {
-            foo(bar: UserType): Boolean @mock
+            foo(bar: AOrB): Boolean @mock
         }
         ';
 
         $this->typeRegistry->register(
-            new LaravelEnumType(UserType::class)
+            new LaravelEnumType(AOrB::class)
         );
 
         $this->mockResolver()
             ->with(null, new Callback(function (array $args): bool {
-                return $args['bar'] instanceof UserType;
+                return $args['bar'] instanceof AOrB;
             }));
 
-        $this->graphQL('
+        $this->graphQL(/** @lang GraphQL */ '
         {
-            foo(bar: Administrator)
+            foo(bar: A)
         }
         ');
     }