+++ CHANGELOG.md
@@ -26,6 +26,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
   get the arguments passed through a client directive https://github.com/nuwave/lighthouse/pull/1184
 - Add `streamGraphQL()` helper method to `\Nuwave\Lighthouse\Testing\MakesGraphQLRequests` for
   simple testing of streamed responses, such as `@defer` https://github.com/nuwave/lighthouse/pull/1184
+- Add support to extend `input`, `interface` and `enum` types https://github.com/nuwave/lighthouse/pull/1203
 
 ### Fixed